# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Q-Wash API
  description: >
    Car-washing queue API: phone + SMS-OTP auth, washing point / service
    CRUD, capacity-aware booking availability, a live queue board, and
    booking history. See `docs/API.md` in the repo for the full prose
    contract and `docs/DATA_MODEL.md` for the schema and algorithms this
    spec is generated alongside.
  version: "1.0.0"
servers:
  - url: /
    description: Same-origin (this server also hosts the docs page)
...
security: []
paths:
...
  /api/v1/washing-points/{id}/photos:
    parameters:
      - $ref: "#/components/parameters/washingPointId"
    get:
      tags: [Photos]
      operationId: listPhotos
      summary: List a washing point's photos
      description: Public. Ordered by sort_order then created_at.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PhotoList" }
    post:
      tags: [Photos]
      operationId: uploadPhoto
      summary: Upload a photo
      description: >
        Requires role staff or admin, scoped to their own washing point for
        staff (404 for another point, same as a nonexistent id). The
        uploaded bytes are content-sniffed server-side; only JPEG/PNG/GIF/
        WebP are accepted (image/svg+xml is deliberately rejected — an SVG
        can embed a script, which would be a stored-XSS vector if served
        back as-is). The first photo for a point always becomes the cover
        regardless of is_cover; a later is_cover=true unsets any existing
        cover first.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file: { type: string, format: binary }
                is_cover: { type: string, enum: ["true", "false"], default: "false" }
              required: [file]
      responses:
        "201":
          description: Uploaded
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Photo" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "413": { description: File exceeds the 10MB limit }
  /api/v1/washing-points/{id}/photos/{photoId}:
    parameters:
      - $ref: "#/components/parameters/washingPointId"
      - name: photoId
        in: path
        required: true
        schema: { type: string, format: uuid }
    patch:
      tags: [Photos]
      operationId: updatePhoto
      summary: Set cover / reorder a photo
      description: >
        Requires role staff or admin, scoped to their own washing point for
        staff. is_cover=true promotes it, unsetting the previous cover;
        there's no way to unset the sole cover directly.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PhotoUpdate" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Photo" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [Photos]
      operationId: deletePhoto
      summary: Delete a photo
      description: >
        Requires role staff or admin, scoped to their own washing point for
        staff. If the deleted photo was the cover and others remain, the
        oldest remaining one is auto-promoted.
      security: [{ bearerAuth: [] }]
      responses:
        "204": { description: Deleted }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /api/v1/owners:
    get:
      tags: [Owners]
      operationId: listOwners
      summary: List owners
      description: Requires role admin. Ordered by name.
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OwnerList" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      tags: [Owners]
      operationId: createOwner
      summary: Create an owner
      description: Requires role admin.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/OwnerCreate" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Owner" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/v1/owners/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [Owners]
      operationId: getOwner
      summary: Get an owner
      description: Requires role admin.
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Owner" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      tags: [Owners]
      operationId: updateOwner
      summary: Partially update an owner
      description: Requires role admin. Any subset of the create fields.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/OwnerUpdate" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Owner" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Access token from OTP verify/refresh, e.g. `Authorization: Bearer <token>`"
  parameters:
    washingPointId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: Missing, malformed, or invalid/expired credentials
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: Authenticated but not permitted
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Not found (or hidden as not-found to avoid leaking existence)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string, example: invalid_name }
            message: { type: string, example: "name is required (max 255 chars)" }
          required: [code, message]
      required: [error]
...
    Owner:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        contact_name: { type: string, nullable: true }
        contact_phone: { type: string, nullable: true }
        contact_email: { type: string, nullable: true }
      required: [id, name]
    OwnerList:
      type: object
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/Owner" } }
      required: [items]
    OwnerCreate:
      type: object
      properties:
        name: { type: string, maxLength: 255 }
        contact_name: { type: string, maxLength: 255 }
        contact_phone: { type: string, maxLength: 32 }
        contact_email: { type: string, maxLength: 255 }
      required: [name]
    OwnerUpdate:
      type: object
      properties:
        name: { type: string, maxLength: 255 }
        contact_name: { type: string, maxLength: 255 }
        contact_phone: { type: string, maxLength: 32 }
        contact_email: { type: string, maxLength: 255 }
...
    Photo:
      type: object
      properties:
        id: { type: string, format: uuid }
        url: { type: string }
        is_cover: { type: boolean }
        sort_order: { type: integer, minimum: 0 }
      required: [id, url, is_cover, sort_order]
    PhotoList:
      type: object
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/Photo" } }
      required: [items]
    PhotoUpdate:
      type: object
      properties:
        is_cover: { type: boolean }
        sort_order: { type: integer, minimum: 0 }
...